Add the ability to specify execution context type
Thread-safety problem: after sending a message (e.g. register)
it is possible to receive response (registered) before corresponding
request is added to internal container (like m_register_requests).
This can happen if, for example, both client and the router are in the
same process and share the same thread pool or io_context.
With the ability to specify the type of exection context it is possible
to use boost::asio::strand to avoid synchronization problems.